requirements.txt
feedparser==6.0.11
notion-client==2.2.1
python-dateutil==2.8.2
requests==2.31.0